move main and change some imports to relative imports
