+++ main.py
@@ -1,5 +1,5 @@
-from fee import Fee
-from order import Order
+from src.fee import Fee
+from src.order import Order
 
 
 def cow_swap():
